// lib/types.ts
export type Job_Entry = {
  uid: string
  modified: string
  company_name: string 
  job_title: string | null
  apply_link: string | null
  job_type_time: string | null
  is_rolling: boolean | null
  location: string | null
  DDL: string | null
  start_time: string | null
  duration: string | null
  salary: string | null
  description: string | null
  keywords: string | null
}